requirements.txt
crewai>=0.30.0
streamlit>=1.32.0
chromadb>=0.5.0
groq>=0.5.0
langchain
langchain-community
pytest>=8.0.0
python-dotenv>=1.0.0
pydantic>=2.6.0
loguru
tqdm
langchain-groq
crewai[litellm]
fastapi
sentence-transfomers